requirements.txt
anyio==3.6.1
attrs==22.1.0
bcrypt==4.0.0
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
click==8.1.3
crashtest==0.3.1
cryptography==38.0.1
distlib==0.3.6
docopt==0.6.2
dulwich==0.20.46
ecdsa==0.18.0
fastapi==0.85.0
filelock==3.8.0
greenlet==1.1.3
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.9.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.16.0
keyring==23.9.3
lockfile==0.12.2
MarkupSafe==2.1.1
more-itertools==8.14.0
msgpack==1.0.4
packaging==21.3
passlib==1.7.4
pexpect==4.8.0
pipreqs==0.4.11
# pkg_resources==0.0.0
pkginfo==1.8.3
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
# poetry-core==1.2.0
# poetry-plugin-export==1.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-decouple==3.6
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
SecretStorage==3.3.3
shellingham==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
# tomlkit==0.11.5
typing_extensions==4.3.0
urllib3==1.26.12
uvicorn==0.18.3
virtualenv==20.16.5
webencodings==0.5.1
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.8.1
